# swagger.json
{
    "swagger": "2.0.0",
    "info": {
        "title": "express&swagger-ui",
        "description": "接口联调",
        "version": "1.0.0"
    },
    "produces": [
        "application/json"
    ],
    "paths": {
        "/login": {
            "post": {
                "x-swagger-router-controller": "home",
                "operationId": "index",
                "tags": [
                    "/user"
                ],
                "description": "用户登录",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "type": "object",
                        "required": "true",
                        "description": "JSON body",
                        "schema": {
                            "$ref": "#/definitions/userReqBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/files": {
            "post": {
                "x-swagger-router-controller": "home",
                "operationId": "index",
                "tags": [
                    "/file"
                ],
                "description": "获取文件列表",
                "parameters": [],
                "responses": {}
            }
        },
        "/uploadFile": {
            "post": {
                "x-swagger-router-controller": "home",
                "operationId": "index",
                "tags": [
                    "/file"
                ],
                "description": "文件上传",
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "type": "file",
                        "required": "true"
                    }
                ],
                "responses": {}
            }
        },
        "/generateImage": {
            "post": {
                "x-swagger-router-controller": "home",
                "operationId": "index",
                "tags": [
                    "/image"
                ],
                "description": "AI图片生成(请勿带有屏蔽词)",
                "parameters": [
                    {
                        "name": "describe",
                        "in": "body",
                        "type": "object",
                        "required": "true",
                        "description": "JSON body",
                        "schema": {
                            "$ref": "#/definitions/ExampleRequestBody"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "definitions": {
        "ExampleRequestBody": {
            "type": "object",
            "properties": {
                "describe": {
                    "type": "string",
                    "default": "英文输入关键描述词"
                }
            }
        },
        "userReqBody": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "default": "用户名"
                },
                "email": {
                    "type": "string",
                    "default": "邮箱"
                },
                "password": {
                    "type": "string",
                    "default": "密码"
                }
            }
        }
    }
}